[SCRIPT] Introduce "fast" wheel building process
This is mostly for users that want to quickly generate a Python wheel if they already have a local CMake build directory
configured (e.g. when using ./build_locally.sh or similar).

In those cases, you can use ./build.sh --fast-build and instead of creating a new build directory when calling the
bdist_wheel command, the Python setup will use that CMake directory to generate the C++ Python extensions and then copy
them in the temporary directory used by bdist_wheel.

Do *not* use this to generate official wheels as this is not 100% tested.